fix: 修复文档主题悬浮按钮事件丢失问题
