+++ src/BootstrapBlazor.Shared/Components/Components/ThemeMode.razor
@@ -1,6 +1,7 @@
-﻿@inherits ComponentBase
+﻿@inherits WebSiteModuleComponentBase
+@attribute [JSModuleAutoLoader("Components/ThemeMode.razor.js")]
 
-<div class="btn btn-circle btn-fade btn-theme-mode icon-theme d-flex">
+<div id="@Id" class="btn btn-circle btn-fade btn-theme-mode icon-theme d-flex">
     <i class="@GetLightIconClassString"></i>
     <i class="@GetDarkIconClassString"></i>
 </div>
